Fix modernize-replace-random-shuffle by clang-tidy
The reason for removing random_shuffle in C++17 is, that the iterator
only version is usually depending on std::rand, which is now also
discussed for deprecation, and should be replaced with the classes of
the <random> header, as std::rand is considered harmful. Also the
iterator-only random_shuffle version usually depends on a global state.
The shuffle algorithm is the replacement, and has URBG as its 3rd
parameter.