Fix starrocks env get error : catalog.java [796]
If the meta-directory used by my cluster of the old version is Doris-meta, upgrade to the new version and check whether the metadata directory exists. If it does not, metadata information will be read from the metadata directory of the old version. However, the current code gets the old metadata directory through DORIS_HOME, which causes fe startup failures