some balance bugfixs and optimizations (#748)
* Optimize the disk selection strategy for cloning tablet (#740)

BE directly uses the disk selected by fe according to the strategy instead of random strategy.

* Bugfix: delete redundant colocate tablet from be when balance finish (#739)

After the colocation table is balanced, the redundant colocate
cannot be deleted from be. In previous StarRocks version,
The deletion depends on the version change, and submitting a
new load job can trigger the deletion of the tablet from be

* Optimize the calculation of disk usage for balance (#734)

During the balance procedure, the disk usage percent is DataUsedCapacity/TotalCapacity.
When OtherUsedCapacity occupies a lot of disk space, and the disk usage is relatively high,
continue to import data will always trigger FE to do balance, but the disk may have no space actually.
Usage percent = DataUsedCapacity / (AvailableCapacity + DataUsedCapacity) is more reasonable.