Enhancement: share TabletSchema to reduce memory usage (#846)
* Share TabletSchema to reduce memory usage

Close issue #807

A 64 bits integer field named `id` is added in `TabletSchemaPB`, with its
value assigned by the master FE node, and the following properties
are ensured by the master FE:
- all tablets from the same table or materialized view will have the same id
- tablets from different tables or materialized views will have different id.

On each BE node, a global hash map was created to map from the id to
std::shared_ptr<TabletSchema>. When creating a new tablet or loading
already exists tablets from disk after the restart, the id will be checked
and only one TabletSchema will be created for all tablets with the same
id.

Please note that this patch only works for newly-created tablets. For those
already-created tablets, their TabletShemas will not be shared among each other.