Cherry-picked latest bugfix to 1.19 (#1071)
* BugFix : Correct batching partition semantics at week interval when the week span two years (#654)

1. The problem is to make batching partition semantics is consistent with dynamic partition.
In china, the Monday is the first day of week. In western country, the Sunday is the first day of week. The semantics is should be consistent between batching partition and dynamic partition. If the option is not set, the Monday will be the first day of week.
2. The 1st January is the first week of every year. Every year have 52 weeks. The last week will end at 31st December.

* Fix batching partition semantics for natural week/year/month. (#997)

Fix #996 .
batch partitions semantics should match dynamic partitions.

* Fix .Net Mysql.Data connect starrocks (#1014)

* Return success when drop not exist materialized view #1026

The use send a DDL statement to drop materialized view like the following.
```
DROP MATERIALIZED VIEW if exists t2;
```
If t2 is not exist, the DDL should success instead of throw an exception.
And also, the related unit test is no longer used and to be removed.