Fix routine load loss data bug (#1074)
The aborted transaction still carries the offset information. The master does not advance the offset if the error reason is not all partitions has no data to load. However, the followers does not judge the logic, it still advance the offset. Unfortunately, the vote procedure is been triggered. The follower change to master and the data is loss. So the pull request is add the logic to check the error reason when advancing the offset.