Fix the dead lock bug of data stream recvr (#1631)
The sender will have have multiple chunks. When reaching the memory limit, sometimes the ChunkQueue will be empty, but PendingClosure is not empty. The get_chunk() will keep waiting forever. The pull request guarantee that the last chunk of one request is consumed, the response will be send to sender.