Load index lazily to reduce the first I/O time of queries (#1579)
When reading column data, all the indexes of column will be loaded, no matter whether they are used or not. But reading index will incur random I/O, which will hurt first-time reading performance.
So we can just load the necessary indexes to improve the first-time reading performance.
I use the Query 2.1 of SSB 100G data set to test performance, it indicate that the latency can be reduced about 200ms of 4300ms when loading zone-map index of three columns.