Remove version_hash in FE code paths.
The version_hash is used to differentiate the different loads with the same version.
After the load transaction is changed to two-phase commit, the different loads
will be guaranteed with different versions. The version_hash is already not used in StarRocks, and can be removed.
This pull request is to remove the version_hash in FE code paths.
To be compatible with the old version, the version_hash used to exhibition in Statement will be remained.