Optimize the return type of some date functions to optimize group by key
1. optimize function for year/month/day/hour/minute/second
2. In order to be compatible with grayscale upgrades, we are not removing the old implementation for now

For some time function, we don't need to optimize the separate year because the to_date function is currently inlined, so some of the seemingly redundant calculations have already been optimized away.

I compared the assembly code of the separate implementation and the current implementation and they are identical.

year: https://godbolt.org/z/d94joTqE5
day: https://godbolt.org/z/h9K7MP97b

We can't verify the effect of this optimization yet, because currently our FE plan is that the return value of year(not-nullable(input)) is nullable, and we need to make FE support this optimization