fix fail to prepare tablet reader bug (#2026)
Now publish version will succeed when quorum succeeds. But the finished transaction will update the version of all replicas to a visible version, which will lead to query the unsuccessful replica and result in query failure. Fix the bug by skipping updating the unsuccessful replica's version when the transaction is finished.