[Vertical compaction] calculate reader chunk size according to memory limit
The default reader chunk size is 4096. When the string column is large,
the chunk read from each segment will be very large and takes up a lot of memory.

So reader chunk size needs to be calculated according to memory limit and column size.

Fix calculation bug when the rowset is overlapping in horizontal compaction.