Wait until memtable flushed to disk if memory limit exceeded (#2188)
Summary: when the loading memory usage exceeds the limit, block the RPC thread
until the memtable flushed to disk or, in other words, stop the sender from sending
more chunks until the memtable has been flushed to disk.

It should be noted that when the memory usage exceeded the limit, a bunch of small
files may be created on disk, this is not a new issue introduced by this patch, but this
patch may increase the possibility. A new patch is needed to solve the problem of a
bunch of small files.