Fix the int overflow bug of DeleteHandler (#2337)
When doing checkpoint, it is to create a separate catalog object, first load the image, then replay the bdb log. So this PR #667 Clearing only the deleteInfo from existing memory does not reduce the number of deleteInfo in the image. Do not add the outdated delete info when replaying bdb log, this will reduce the number of deleteInfo in the image when doing checkpoint.