Make error message of load procedure more concise
The previous error message of load procedure is misunderstanding.
What cause the failure cannot be seen at a glance.
This pull request want to clarify several types of errors.

1. column count does not mismatch
```
Value count does not match column count. Expect $0, but got $1.
```

2. Value overflow
```
Value '$0' is out of range. The type of '$1' is '$2'.
```
$0 refers to the input value.
$1 refers to column name.

3. Decimal overflow
```
Decimal '$0' is out of range. The type of '$1' is '$2'.
```
$0 refers to the input decimal.
$1 refers to column name.

3. Type incompatibility
```
NULL value in non-nullable column '$0'.
```
$0 refers to column name.

4. String length exceeds the limit
```
String '$0'(length=$1) is too long. The max length of '$2' is $3.
```
$0 print up to 100 characters. If the length exceed 100, use ellipsis instead.
$0 refers to the input string
$2 refers to column name.