Add setBadForce property to distinguish between manual setting to bad and automatic system setting to bad (#2480)
The status of replica, after being manually set to bad, will be reset to ok after tabletReport. But the reality is that the tablet on be is corrupted but not detected. The current handling is a bit tricky: replica status is not reset to ok in master, but the updateStatus log is recorded. So that the bad replica can still be recovered by copy. We should add another property to distinguish between manual setting to bad and automatic system setting to bad. Manually setting to bad can not be reset to ok after tabletReport.