Add helper class ColumnVisitorAdapter
Summary: Add a helper class, ColumnVisitorAdapter, to simplify the implementation of a new
ColumnVisitor without writing many tedious codes for each specific column type.

By inheriting the ColumnVisitorAdapter, the derived classes can be implemented with template
functions, so there is no need to implement a visit method for each of the specific column type.