BugFix : FE should clear txn about delete after restart (#2752)
The transaction type of Delete is FRONTEND. 
For the FRONTEND type, only asynchronous operations need to write logs, that is, 
```
editLog.logInsertTransactionState(transactionState)
```
in DatabaseTransactionMgr.java.

Now DELETE is a synchronous operation. 
If FE writes this log and restarts during delete, the transaction state will exist until the transaction timeout.

So I add a new type for mark delete to solve this problem.