hive metadata cache of fe automatic incremental synchronization (#3036)
Support automatic synchronization of metadata updates in hive metastore(hms).

Each fe node periodically obtains the hms incremental update event, parses it, and updates the cache in fe to achieve automatic synchronization. The hms interface we need to call has getNextNotification and getCurrentNotificationEventId.

A metastore event is a instance of the class org.apache.hadoop.hive.metastore.api.NotificationEvent. Metastore can be
configured, to work with listeners which are called on various DDL operations like create/alter/drop operations on database, table, partition etc. Each event has a unique incremental id and the generated events are be fetched from Metastore to get incremental updates to the metadata stored in Hive metastore using the the public API get_next_notification. These events could be generated by external metastore clients like Apache Hive or Apache Spark configured to talk with the same metastore.

We could process the events that fetched during each rotation according to the following points.

1. create batch event
   Create a batch of events based on the partition or table. It is convenient for us to process batch event in parallel.
2. filter event
    For some partitions that are not cached in fe, and some events that only modify transient_lastDdlTime, we ignore these 
    events.
3. merge event
    For multiple alter table/ alter parititon events of a table or partition, we only process the last one.
4. parallel process
    Because batch events are partition-level, each batch event is processed in parallel.

Note that this feature is still a work in progress and additional improvements will be done in subsequent patches.