Make APIs of RandomAccessFile similar to those of io::RandomAccessFile
The final goal is to delete the class RandomAccessFile defined
in env.h and leave only the class io::RandomAccessFile defined
in io/random_access_file.h. To achieve this, the first step is
making those two classes have similar APIs, and the next step
would be replacing the usage of the first class with the second one.