Fix bug CLONE is unsuccessful duo to different storage medium during … (#3763)
This is due to the default value of FE's configuration enable_strict_storage_medium_check is false.
But this parameter is not considered when doing clone .
So when enable_strict_storage_medium_check is true. the same logic as before.
if enable_strict_storage_medium_check is false.
We will choose the same medium first, and if it cannot be selected, then choose those mediums that were ignored before.