Make APIs of RandomAccessFile similar to those of io::RandomAccessFile (#3565)
- Redefine `Status RandomAccessFile::read(Slice* result)` as `StatusOr<int64_t> RandomAccessFile::read(void* data, int64_t size)`;
- Redefine `Status RandomAccessFile::read_at(uint64_t off, const Slice& result)` as `StatusOr RandomAccessFile::read_at_fully(int64_t off, void* data, int64_t size)`;
- Redefine `Status SequentialFile::read(Slice* res)` as `StatusOr<int64_t> SequentialFile::read(void* data, int64_t size)`;

The final goal is to delete the class RandomAccessFile defined
in env.h and leave only the class io::RandomAccessFile defined
in io/random_access_file.h. To achieve this, the first step is
making those two classes have similar APIs, and the next step
would be replacing the usage of the first class with the second one.