BugFix: fix the wrong checksum between different replicas (#3516)
StarRocks uses the CRC algorithm to get the checksums to check the consistency between replicas.
There are two problems. Firstly, the size of chunks between different replicas is not the same.
Secondly, the order of data is not the same because of the out-of-step compaction procedures.

To solve the problem, I change the checksum algorithm to XOR. The result will be stored as a 64-bit integer.
Since the data between replicas is not the same, the checksum will reflect this problem, and the error probability is 1/(2^64).
XOR also obeys the commutative law. I can calculate the checksum Rowset by Rowset. XOR will improve performance and reduce memory consumption.

Using the lineorder table(TPC-H), I test it with 100 million rows and 20 rowsets.
The test indicates that the memory consumption reduced from 500MB to 140MB. And the time is reduced from 24s to 9s.
As the number of rowsets grows, the performance improvement effect will be more obvious.