Create BlockManager dynamically based on URI schema (#4158)
This modification will be used for the future implementation of decoupling storage and computing.

This modification will NOT affect the process of data loading and reading.

Instead of using a global-shared instance of BlockManager, create BlockManager dynamically based
on the scheme of a path or URI.

By doing this we can support saving segment files on remote storage by changing the value of Tablet::schema_hash_path(). For example, if Tablet::schema_hash_path() returns a path
prefixed with s3://, the segment files will be stored on the remote S3 storage automatically.