Add horizontal and vertical compaction tasks for event_based compaction framework (#4083)
The original compaction framework polled the candidate tablets one by one. The performance is low, and the overhead is high. Now I change the mechanism of picking compaction candidates from polling to triggering by the event(#3311).
Considering StarRocks split the columns of a table into groups, it can improve compaction performance (#2189). This pull request adds the horizontal and vertical compaction to the event_based compaction framework.