Add compaction policy files (#4086)
Add BaseAndCumulativeCompactionPolicy for new compaction execution framework. BaseAndCumulativeCompactionPolicy is created based on CompactionContext to decide whether the tablet should be compacted, which type of compaction should be done, and how many rowsets should be compacted. And BaseAndCumulativeCompactionPolicy uses the same strategy as before. The details are as follows: when the compaction score is above a threshold, it is considered to be a compaction candidate. It uses the same strategy to pick rowsets as before.