[cherry-pick] [branch-2.0] BugFix: Memory leak of rowset_id and stack overflow in copy_file (#4392)
* BugFix: stack overflow in copy_file (#4347)

The stack memory limit set by the operating system can be changed using `ulimit -s `, and allocating memory greater or equal to the stack memory limit will cause be crash. If the stack memory limit is no more than 1MB, be will crash because try to allocate 1MB stack memory in function copy_files. We should not allocate large memory in the stack; allocate memory from the heap to solve this problem. This code is duplicated with function copy_file in utils/file_utils.cpp, merge the two copy_file function in utils/file_utils.cpp and storage/utils.cpp

* Rowset id cache will leak in some scenarios:
   1. allocate rowset id success, but create rowset writer failed
   2. create rowset writer success, but no data written
      ...
Anyway, if allocate rowset id success, but does not generate rowset_ptr, we don't release the rowset_id, resulting in a memory leak.

In addition, when we do load data, we create a delta_writer for each tablet in advance, but only a few of the delta_writer may have data written to them. This speeds up memory leak and output a lot of useless rollback logs in be.INFO

This pr will call release_rowset_id in deconstruct function of delta_writer if rowset is not built and reduce the rollback log.