The fe node get crashed when receive beginRemoteTxn request if it is … (#4418) (#4585)
* The fe node get crashed when receive beginRemoteTxn request if it is not leader.

===========cause=================
When an external table is created and the host property set to a
follower/observer node of the source cluster. If we insert some values
into this external table, the target fe node will receive
beginRemoteTxn rpc, and get crashed because of wal writing
is needed for beginRemoteTxn, which is forbidden in non-leader node.

==========solution===============
For non-leader fe node, it will forward beginRemoteTxn/commitRemoteTxn/
abortRemoteTxn to leader node and return back response from leader.

* return failed response rather than throws exception if forward request failed

Co-authored-by: wulei <87756338+wuleistarrocks@users.noreply.github.com>