Remove some unnecessary inline keywords
1. A member function defined in a class definition is implicitly
inline, so there is no need to put the inline keyword in this case.
2. The decision of whether a function is or is not inline is an
implementation detail that does not change the observable semantics
(the “meaning”) of a call. Therefore the inline keyword should not
go within the class’s public: (or the protected: or private:) part,
so it needs to go next to the function’s definition.

References:
https://llvm.org/docs/CodingStandards.html#don-t-use-inline-when-defining-a-function-in-a-class-definition
https://isocpp.org/wiki/faq/inline-functions#where-to-put-inline-keyword