[optimize] Reducing the memory footprint of columnReader. (#4873) (#4875)
In a large wide table scenario, when there are many columns, a large number of Column Readers will occupy a lot of memory. This pr is mainly to strip the data structures that can be shared between Column Readers and put them into Segment.
Share these data field between ColumnReaders: