Fix invalid migration for tablets (#4993)
When clone tablet from path A to path B, and then back to path A,
it will fail because the original tablet remains in path A. The
original tablet is GCed async every one hour. Fix it by removing
the shutdown tablet when cloning. Add ut to test this case.