Fix bdb RollbackException bug (#4978)
In bdb, after transferring the master, there may be logs in the old master that have not been synchronized to other nodes, and these logs need to be cleaned up. After cleaning these logs, the bdb environment will throw a RollbackException, and the environment should be rebuilt. So we should catch the RollbackException, and rebuild the bdb environment.

The current handling of InsufficientLogException is somewhat problematic; we should rebuild the bdb environment after synchronizing data from other nodes.

Since we handle the InsufficientLogException in the getDatabaseNames() function, there is no need to address the exception in the open() function, so remove the extra code.