Fix fail to prepare tablet reader bug
When publish stuck in one replica and last failed version is > 0,
the replica will not be treated as error replica, and FE will update
the replica's version to the newest visible version of partition.
and the query will failed because missing version because the publish
version do not finish in BE, the version is actually not querable.