Backup restore compatible with apache doris
Backup table from apache doris 0.14, and restore to StarRocks will throw error.
Because tabletsOrder is stored as long in apache doris and is parsed in StarRocks as String.

In addition, |tablets| is a map in backup __info_<timestamp>.<md5> file, so tablet order is not guaranteed.

So we use Long in tabletsOrder for compatibility with apache doris,
and serialize tabletsOrder to guarante the correct tablet order.