BugFix: consumer not been scheduled when consumers of group larger than three (#5132)
The BE uses a thread pool to execute the consumers of the consumer group, while the size of the thread pool is three. When the number of consumers is larger than 3, some consumers have to wait for the availability of the thread pool. The delay of execution may lead to the task running time beyond routine_load_task_consume_second, and then the hanging consumer would be canceled before being executed.