Build minizip in thirdparty (#5315)
Currently, StarRocks copy minizip codes in self reop. However it is not
necessary, because minizip is included in libz. So we only need to build
it in libz, then we can use it as third-party depedency.
This is the first PR to add minizip in third-part build script. After
this PR, I will remove code in be/src/util/minizip and make StarRocks
depend minizip in libz.