[BugFix] json stream load failed with Transfer-Encoding: chunked set (#5330)
When `Transfer-Encoding: chunked` is set, the content of the HTTP body cannot be known from the HTTP header `Content-Length`.
However, the previous implement of `StreamLoadPipe` need the http body size for initializing `_total_length`, and the `StreamLoadPipe:read_one_message` also needs `_total_length`.

This PR includes the following modifications:
1. remove the `StreamLoadPipe`'s dependency on HTTP body size as the initial size
2. redefine the semantics of `StreamLoadPipe:read_one_message` as `returns data that is written by append in one time`
3. new unit tests for `StreamLoadPipe:read_one_message`

The `StreamLoadPipe` is a pipe to store data from stream load. Every time we use `StreamLoadPipe::append` for JSON, we must append a complete JSON payload. So, it's OK to parse the data as JSON returned by  `StreamLoadPipe:read_one_message`.