[Enhancement] construct row in json object order when column is sparse (#5357)
The previous implement constructed a row from JSON in column order. The JSON scanner iterates the column descriptor and looks up the column name in the JSON object. When the number of columns is much more than the JSON field, there're many useless looking up of the JSON object. The new implementation uses a smarter strategy to construct the columns: If the number of columns is much more (2x) than the number of object fields, lookup the object field key in the dictionary of the column. Otherwise, look up the column name in the JSON object.

## Benchmark
load type: Stream Load
source file: 65MB in total, json array format, 6KB / object, 64 fields /object
table: 330 columns, 10 buckets
| version | throughput |
| -- | -- |
| main | 2.5MB/s|
| c8d00b2e4@perf/sparse-column| 18.2MB/s |