Opt: set size of connection pool of jprotobuf to 16 by default (#5478)
The size of shared connection pool of jprotobuf is 1, which means protobuf RPC requests can only send sequentially. And the timeout for getting connection from the connection pool is 10s.
There are too many RPC requests for 1000-concurrency test. As a result, some requests timeout and add this BE to blacklist.

Modification

* Rename Config.brpc_number_of_concurrent_requests_processed to brpc_connection_pool_size, default is 16, to avoid * the existed users have already change brpc_number_of_concurrent_requests_processed in config file.
* Set the size of shared connection pool to Config.brpc_connection_pool_size.
* set MaxIdleSize to the size of shared connection pool for both shared and non-shared connection pool.
* After the RPC request sending, the connection will be closed, if the number of total connections is greater than MaxIdleSize. Therefore, MaxIdleSize shouldn't less than MaxTotal.
* Remove the branch enable_brpc_share_channel=false