BugFix: BE start failed when enable StroagePageCache (#5593)
When we start BE, we will create `apply thread` for PrimaryKey table to do `rowset apply`.

The `apply thread` start in `load` phase, when the StoragePageCache has not yet been created. So if  `disable_storage_page_cache` is false, the `apply thread` will try to get page from StroagePageCache which will cause a nullptr exception. 

The solution is create StoragePageCache before start `apply thread`.

In addition, the `min_buffer_size` configuration does not seem to be used anymore, delete it.