[BugFix] Fix multi edge order bug by refactor version_graph using vertex directly instead of vertex index (#5621)
The previous implementation used _version_graph, _version_index_map as two structures to implement the adjacency graph. Vertex edge was an index of vertex, not vertex. Using set sort will make a mistake.
In this case, it is unnecessary to use the two structures to implement the adjacency graph. The benefits of compact memory are not comparable to the cost of sorting through the vertex index each time.