[Enhancement] simple password validation & password reuse check (#5637)
This PR is a temporary implementation of password validation & password reuse check. The rules are mainly hardcoded and cannot be fully configurable except for a switch to enable it. We will refactor the whole user privilege framework later to fix this ultimately.

Set enable_validate_password to true to validate the password by these rules:
a. The length of the password should be no less than 8.
b. The password should contain at least one digit, one lowercase letter, one uppercase letter
Set enable_password_reuse to false to prevent changing the password to the previous one.