[Enhancement] Optimize HeapMergeIterator read speed (#5832)
The merge_iterator use heap to order the chunk, each time it will pop a min_chunk from the heap.
(1) If the min_chunk is less than other chunks in the heap, then it will swap the whole min_chunk out.
(2) If not, it will read only one line from the min_chunk and append it to the result chunk, then push this min_chunk back to the heap.

The above (2) procedure is the bottleneck of the merge_iterator. Each time we process one line, we need to append this line, and pop and then push the min chunk to the heap.

To solve this problem, in each time, i will get the last row in min chunk whose key value is smaller than other chunks in heap, and then append all rows together to the result chunk.

In my test, there are 160 million random data. The horizontal base compaction before optimization takes 155 seconds, after optimization it only takes 107 seconds, improve by 1.4486x.

Signed-off-by: xyz <a997647204@gmail.com>