[enhancement] Opt the binary search of find partition end in Analytor (#5864)
In the past, every time a new Chunk was read, the start pos of the binary search was PartitionStart. Currently, we execute binary search from the last pos, to reduce the frequency of compare.

This is not the main point that affects the calculation performance of the window function. The actual test has no effect on the performance. The main purpose is to release the memory earlier of the PartitionColumn in advance later.