[BugFix] Remove limit for committing empty transaction (#5768)
The previous implement forbids empty transactions and would return the error `all partitions have no load data`. 

But when loaded data are filtered by `JSON path` or `where condition`, the JSON/CSV scanner may return zero rows and the transaction would be aborted with the error `all partitions have no load data`. The result is confusing for the user. Besides, in routine load, the committed offset would not be advanced.

For compatible consideration, `empty_load_as_error` is added to the FE config. The default behavior stays the same as before.