[cherry-pick][branch-2.2] some bugfix (#5891)
* [Feature] add some metrics for pipeline (#5847)

* [Bug] Fix memory Leak in Pipline when prepare failed (#5882)

When prepare stage failed, pipeline driver won't be added in
driver_exec_queue. but query_ctx or fragment_ctx has been created. So
they won't clean.

In this patch, we add a function `_fail_cleanup` which could help
cleanup query_ctx and fragment_ctx.

In addition to that I have reduced the dependency of executor to
fragment_ctx. If prepare fails, ctx will be automatically destructured,
because fragment_ctx was created with a unique_ptr which will cause
executor hold a invalid pointer.