[BugFix] close segment file when rowset writer meets error
When beta_rowset_writer meets error, it will not close the segments file it previously opens.
In the destructor, we will try to delete the segments file which are still opened and raises an error.

In this pr, i will close the segment file which is going to be deleted for horizontal writer.
For vertical beta rowset writer, we directly use wbblock abort api to close and delete the file.

Also, for horizontal writer with primary key, the normal segment file number before final merge is 0,
we should skip the delete when final merge not starts.

Signed-off-by: xyz <a997647204@gmail.com>