[BugFix] Add retry for fe thrift rpc (#5656)
In some cases, such as fe restart, a broken conn will be returned from the connection pool, because there is no validation of the conn before being returned to user. We should reopen the conn and retry the rpc.
But for the timeout exception, it may be a network timeout or the target server may be running slow. We just reopen the conn in this case but do not retry.
Since we do not retry for the timeout exception, there is no need to distinguish between the types of statements that should be retried, so we discard the isQuery param from MasterOpExecutor.