[Bugfix] delay context switch to next time if _overflow_read_chunk is not empty (#5991)
This pr fixes the second bug introduced by StarRocks#5701

In the current implementation, we will perform a context switch even if the _overflow_read_chunk
is not empty, leading to some data loss. We need to wait until the data in _overflow_read_chunk
is processed in the next batch to perform a context switch safely.