[Enhancement]PrivTable list->map to speed up matching (#5945)
This PR implements Plan 1 of #5944.

This commit changes the data structure of PrivTable from List to Map, using UserIdentity as the key because after
identification of the UserIdentity of the context is an existing entry and requires no fuzzy matching.
The authorization time will significantly improve with this map.

I added a test case to see the performance change. Assuming there are 500 users who each have the privilege of 250 tables, the time of authorization speed up to nearly 100 times(10min+ -> 7s) after this PR.