[Enhancement] Disable enable_exchange_pass_through default to false (#6055) (#6063)
Currently, if enable_exchange_pass_through is turned on. The performance has no improve on benchmark test, and it will cause memory statistics problem of fragment instance, It also will introduce the problem of cross-thread memory allocate and release. So i temporarily disable the enable_exchange_pass_throuth. I will turn on it after all the above problems are solved.