[BugFix] Json loader will lead to BE crash if it reads an error row. (#6061)
Now, the json loader will lead to BE crash if it meets an error row.
In the non strict model, the data(number, string) which cannot be parsed by simdjson will be filled with NULL. However, in the strict model, the data which can not be parsed by simdjson will not be filled with NULL. This will lead to the error column containing one row less than other columns. In this pr, we will remove the error row before we continue to parse other rows.

Apart from this, there's a bug in simdjson which will recognize -9223372036854775808 as unsigned number, this pr also adds additional logic to handle this.