[Enhancement] add one interface for conservative estimize the size of one row (#6333)
At present, when doing `SortedSchemaChange`, the length of the varchar is used to estimate the size of one row, but when users use `String (Varchar(65536))` a lot, the estimation will be inaccurate, so the `SchemaChange` task that will execute failed because of invalid memory estimation.

The new strategy is to use a conservative estimate for the first chunk and then calibrate row_size every time a chunk is read. Whether to perform Sort and Flush is determined according to the current actual memory + the expected memory of the next chunk.