[UDF] optimize UDAF group by performance (#6082)
In the past, we used to hold a global ref directly for the State of the aggregation function. But this has led to some problems:

it is easier to trigger GC
need to call the SetObjectArray method before executing batch_update, which causes a lot of JNI overhead
this PR uses an optimization like this
Instead of holding a global ref handle directly, each aggregate function holds a FunctionStates (a list), and the UDAF State uses an int to represent a handle. (This imposes some restrictions, such as the number of group by keys cannot exceed int32, but in the old implementation framework it is usually almost impossible to execute when hundreds of thousands of global refs exist at the same time)