[Enhancement] Ignore the expired txns and jobs when replay (#6417)
1. Discard expired transaction/export job/load job/delete info when replaying & loading image.
2. Refactor: use treemap instead of list + in-memory sorting. 

To verify this PR, I created a cluster that won't make checkpoint or clean jobs for a long time by modifying many configurations, such as extending the time of checkpoint interval and extending the expiration time of the transaction. Then I created a primary key table and executed 10k delete & insert SQL, which ended with 70k journal logs. Finally, I modify all those configurations to an expected value and restart FE. It turns out that the memory usage of the master FE magnifically decreases after this PR. 

1. The virtual memory of FE after replaying all journals is 18.7g in the old version and 12.7g after this PR.
2. The virtual memory of FE after making the checkpoint is 23.1g in the old version and 13.8g after this PR.
3. RSS of FE after replaying all journals is 1.1g in the old version and 690M after this PR.
4. RSS of FE after making the checkpoint is 1.4g in the old version and 877M after this PR.