[BugFix] fix query memory usage metrics are negative after query close GRF (#6662)
If the memory in the GRF is requested from the query level and is not freed at this time. When the query-leverl tracker is destructured, we will return the unrecovered values to the memory in the query-pool statistics.
If we rebind the memtracker in the query-pool, it will be destructured twice at this time

(cherry picked from commit be5bc71f18824ef322b83e01bf69570e81c0f028)