[Enhancement] Enable memtable sync write when there are too many memtables in the queue
When the concurrent stream load is fast enough, memtable write and flush thread can not catch up, continue to write the memtable will only make the memtable queue even longer, resulting in a large number of memtable in the memory.
In order to solve this problem, we can transfer from the async memtable table write to sync memtable write. Let the import speed to be slowed down and flush the memtable table in the queue.

Signed-off-by: xyz <a997647204@gmail.com>