Intersection selection generated by GRFs instead of the last GRF wins  in GRF exploration phase. (#4934) (#6817)
When runtime-bloom filters are evaluated on chunks, it works in a style of exploring 1 chunk and exploiting next 31 chunks.
In exploration phases, all filters are evaluated and selectivity are computed, only the filters of high selectivity are chosen to be applied to chunks in exploitation phases. selections generated by filters should be intersectioned, but mistakenly, the last filter(that has largest filter id) overwrite the shared selection, if the last filter has a poor selectivity(a filter in tpcds q85 is just this case, selectivity >0.97), then the chunk is almostly not filtered.