[Feature] Compression Context Pool
Previously, each time we perform an LZ4F or ZSTD compression, we need to create a LZ4F or ZSTD compression context. Each compression context's memory size is not small, they all allocate from the central cache of tcmalloc, not local thread cache. For tcmlloac, allocating memory from central cache needs to acquire the spinlock first. So when many threads create compression context together, they will compete to get the spinlock of tcmalloc's central cache, which will lead to the degration of total perform.

In addition to the problem above, previsous compression will first allocate very large buffer(equal to max_compressed_len) for compression output, and then shrink it after compression. Allocating and freeing large blocks of memory is done throught the central cache of tcmalloc, which may also leads to a lot of conflict.

For the first problem, i introduce a compression context pool for LZ4F and ZSTD, before compression we will request a compression context from the pool, after compression we will return the context to the pool, in this way we can reuse the same context many times.

For the second problem, i adds a compression buffer in each compression context, this compression buffer is shared under the same compression context, the compression result will first be saved on the compression buffer, and then copy to the target compression body. In this way, we can avoid to allocate a very large buffer and then shrink it.

Signed-off-by: xyz <a997647204@gmail.com>