Fix UT RoutineLoadManagerTest.testExpired 2 (#6908)
This UT failed because the time setting was not good. The time was changed after the metadata was saved, and there was a logic to re-read the metadata. At this time, the previously set time was invalid, which caused the UT to always fail. so. Just put this setup time before the save element.