[Enhancement] Optimize task query metadata when there are many dbs (#6893)
We can first send an RPC to count which DBs all Task/TaskRun contains and then reduce the RPC query for each DB.
Usually, users do not build many DBs and send Tasks in different DBs simultaneously.
It can be further optimized later, and the indexes of which DBs are included in the Task are built in advance. At present, no deep optimization is performed.