[Enhancement] Report tablet version after publish version task finish (#7123)
Fixes #6391

Problem Summary(Required) ：
Currently, BE's publish_version task only reports whether it is successful or not but does not contain which version the replica has and which versions are missing. This will limit FE's ability to publish txn or schedule clone tasks.

For example, if a replica has version 1 2 4 5, and publish_version 3 succeeds, the version should grow to 5, and txns waiting for version 4, 5 to be published can proceed.

This PR also does a major refactor to the PublishVersionTask execution-related code.