[Refactor] Prepare cleanup SchemaChangeJob and RollupJob (#6947)
To remove SchemaChangeJob and RollupJob, we need two step.
Step 1:
For BDB Image , we can only write 0 for OldAlterJob for compatibility to cleanup the image in 2.4 version.
For BDB Log, we just remove this id, and user should config ignore_unknown_log_id=true to ignore this log.
Add mark @ Deprecated for some class and OP_ID.
Step 2:
Remove all relevant code, keeping the smallest part of the code. in 2.5 version.