[BugFix] choose wrong compaction algorithm in some case
When performing base compaction on the tablet with the following rowsets:

   "rowsets": [
        "[0-175183] 11 DATA NONOVERLAPPING",
        "[175184-175232] 0 DATA NONOVERLAPPING",
        "[175233-175278] 0 DATA NONOVERLAPPING",
        "[175279-175327] 0 DATA NONOVERLAPPING",
        "[175328-175369] 0 DATA NONOVERLAPPING"
    ]

Because there are 11 segments and columns num are more than 5, so it will choose the vertical compaction.
When creating tablet reader, it will merge 11 segment iterator into one union iterator, and create a HeapMaskIterator with only one union_iterator.
In the `new_heap_merge_iterator`, it will return the iterator immediately if the input iterators len is 1.
So here we will then call the get_next(Chunk* chunk, std::vector<RowSourceMask>* source_masks) on the union_iterator,
However, the union_iterator doesn't implement get_next(Chunk* chunk, std::vector<RowSourceMask>* source_masks), and
we get the warning `get chunk with sources not supported`.

In fact, we should count the number of segment iterator after creating tablet reader, only one union iterator in this case,
and choose the horizontal compaction.

Signed-off-by: xyz <a997647204@gmail.com>